\name{CNVRanges}
\docType{data}
\alias{CNVRanges}
\title{Genomic locations and indices of the simulated CNVs.}
\description{
      This data set gives the starts, ends, and the integer copy number of the
      simulated CNVs in the data set \code{\link{XRanges}} object.}
\usage{CNVRanges}
\format{A GRanges object with 20 rows and 40 value columns across 1 space.}
\source{http://www.bioinf.jku.at/cnmops/cnmops.html.}
\references{Guenter Klambauer, Karin Schwarzbauer, Andreas Mayr, Djork-Arne Clevert, 
Andreas Mitterecker, Ulrich Bodenhofer, Sepp Hochreiter. 
\emph{cn.MOPS: mixture of Poissons for discovering copy number variations in next 
generation sequencing data with a low false discovery rate.}
Nucleic Acids Research 2012 40(9); doi:10.1093/nar/gks003.}
\keyword{datasets}
