% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsAccess.R
\docType{methods}
\name{cnvs,CNVDetectionResult-method}
\alias{cnvs,CNVDetectionResult-method}
\title{This generic function returns CNVs of
 a CNV detection method stored in an instance of 
\code{\link{CNVDetectionResult-class}}.}
\usage{
\S4method{cnvs}{CNVDetectionResult}(object)
}
\arguments{
\item{object}{An instance of "CNVDetectionResult"}
}
\value{
\code{cnvs} returns a  eturns a "GRanges" object containing
the CNVs.
}
\description{
This generic function returns CNVs of
 a CNV detection method stored in an instance of 
\code{\link{CNVDetectionResult-class}}.
}
\examples{
data(cn.mops)
r <- cn.mops(X[1:100,1:5])
cnvs(r)
}
\author{
Guenter Klambauer \email{klambauer@bioinf.jku.at}
}

