% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsAccess.R
\docType{methods}
\name{normalizedData,CNVDetectionResult-method}
\alias{normalizedData,CNVDetectionResult-method}
\title{This generic function returns the normalized data
 of a CNV detection method stored in an instance of 
\code{\link{CNVDetectionResult-class}}.}
\usage{
\S4method{normalizedData}{CNVDetectionResult}(object)
}
\arguments{
\item{object}{An instance of "CNVDetectionResult".}
}
\value{
\code{normalizedData} returns a "GRanges" object containing
the normalized data.
}
\description{
This generic function returns the normalized data
 of a CNV detection method stored in an instance of 
\code{\link{CNVDetectionResult-class}}.
}
\examples{
data(cn.mops)
r <- cn.mops(X[1:100,1:5])
normalizedData(r)
}
\author{
Guenter Klambauer \email{klambauer@bioinf.jku.at}
}

