\name{show}
\docType{methods}
\alias{show-methods}
\alias{show,CNVDetectionResult-method}
\title{Displays the result object of a copy number detection method.}
\description{
  Displays method for S4 class \code{\linkS4class{CNVDetectionResult}}
}
\usage{
\S4method{show}{CNVDetectionResult}(object)
}
\arguments{
  \item{object}{An instance of a "CNVDetectionResult".}
}
\value{
  Displays the result object of a CNV detection method.
}

\author{
  Guenter Klambauer \email{klambauer@bioinf.jku.at}
}

